Prevent deleting a product if it still has a ticket assigned to it (#1940)
* Prevent deleting a product if it still has a ticket assigned to it

* PHP Linting (Pint)

* Fix message

* PHP Linting (Pint)

---------

+++ app/Http/Controllers/ProductController.php
@@ -224,6 +224,12 @@ class ProductController extends Controller
             return Redirect::back();
         }
 
+        if ($product->ticket) {
+            Session::flash('flash_message', 'You cannot delete this product because there is still a ticket associated with it.');
+
+            return Redirect::back();
+        }
+
         $product->delete();
 
         Session::flash('flash_message', 'The product has been deleted.');
